refactor(persona): the work-turn text builders live in persona/work_burst.rs (pure code motion)
held_work_burst, own_recent_thoughts, work_board_anchor — pure text, no I/O — out of the service loop (5,324 → 4,995 lines today).
